requirements.txt
# === Flask Backend ===
Flask==3.1.0
flask-cors==5.0.1
gunicorn==21.2.0  # Lightweight, stable for prod

# === Face Emotion Detection ===
deepface==0.0.93
opencv-python-headless==4.11.0.86  # ✅ headless = no GUI, saves memory
mtcnn==0.1.1
tensorflow==2.16.1  # Latest stable, GPU-friendly
tf-keras==2.16.0

# === Voice Emotion Detection ===
SpeechRecognition==3.10.1  # More stable with Google engine
pydub==0.25.1
librosa==0.11.0
soundfile==0.13.1
audioread==3.0.1

# === Text Sentiment Analysis ===
torch==2.2.2
torchaudio==2.2.2
transformers==4.36.2

# === System + ML Utils ===
numpy==1.26.4
scipy==1.13.1
pandas==2.2.3
joblib==1.4.2

torchvision==0.17.2

# === Backend I/O and Audio ===
ffmpeg-python==0.2.0  # Optional, backend audio conversion (if needed)

uvicorn==0.29.0  # If you want to run async mode or migrate to FastAPI

python-dotenv==1.0.1

openai==1.14.2

gdown


